Menu: extract the Vision tab into build_vision.cpp
Verbatim move of the camera sections (CSI controls, resolution, zoom,
mirror crop, single/multi-cam layouts, USB camera PiPs, Android mirror,
Vision Assist). The overlay placement factories (make_position_items,
make_size_slider, make_rotation_items) were captureless lambdas shared
with the Face Display panel preview, so they become inline functions in
item_factories.h.
