+++ CMakeLists.txt
@@ -207,6 +207,7 @@ set(SOURCES
     src/menu/build_communications.cpp
     src/menu/build_quick.cpp
     src/menu/build_system.cpp
+    src/menu/build_vision.cpp
     src/menu/menu_system.cpp
     src/menu/file_picker.cpp
     src/menu/face_editor.cpp
